follow-up: added more fixes to the domain model detected
Detected during the follow up session
