+++ src/main/java/acme/entities/patronageReport/PatronageReport.java
@@ -1,10 +1,12 @@
-package acme.entities.announcements;
+package acme.entities.patronageReport;
 
 import java.util.Date;
 
 import javax.persistence.Entity;
+import javax.persistence.ManyToOne;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
+import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Past;
@@ -12,6 +14,7 @@ import javax.validation.constraints.Past;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.URL;
 
+import acme.entities.patronage.Patronage;
 import acme.framework.entities.AbstractEntity;
 import lombok.Getter;
 import lombok.Setter;
@@ -19,30 +22,34 @@ import lombok.Setter;
 @Entity
 @Getter
 @Setter
-public class Announcement extends AbstractEntity {
+public class PatronageReport extends AbstractEntity {
 
 	// Serialisation identifier -----------------------------------------------
 
 	protected static final long serialVersionUID = 1L;
-	
+
 	// Attributes -------------------------------------------------------------
 
+	@NotBlank
+	protected String automaticSequenceNumber;
+	
 	@Temporal(TemporalType.TIMESTAMP)
 	@Past
 	@NotNull
 	protected Date creationMoment;
-	
-	@NotBlank()
-	@Length(min=1, max=100)
-	protected String title;
 
 	@NotBlank()
-	@Length(min=1, max=255)
-	protected String body;
-	
-	@NotNull
-	protected Boolean flag;
-	
+	@Length(min = 1, max = 255)
+	protected String memorandum;
+
 	@URL
 	protected String link;
+
+	// Relationships -------------------------------------------------------------
+
+	
+	@Valid
+	@NotNull
+	@ManyToOne(optional=false)
+	protected Patronage patronage;
 }
