+++ src/main/java/acme/features/any/userAccount/AnyUserAccountController.java
@@ -1,4 +1,4 @@
-package acme.features.any.useraccount;
+package acme.features.any.userAccount;
 
 import javax.annotation.PostConstruct;
